Revert "Nguyen empolyee"
