--- src/pages/createUser/index.css
@@ -1,33 +1,0 @@
-#createuser {
-    width: 600px;
-    margin: 90px;
-    padding: 20px;
-    margin-left: 380px;
-    
-  }
- #createuser_btn{
-    left: 485px
-} 
-#createuser_name{
-    width: 472px;
-    left:   23px
-}
-#createuser_username{
-   width: 470px;
-    left:-3px
-}
-#createuser_email{
-    width: 472px;
-    left:27px
-}
-#createuser_phone{
-    width: 472px;
-    left:22px
-}
-#createuser_pin{
-    width: 474px;
-    left:41px
-}
-.mainBody{
-    border-top: 0.01px solid silver;
-}
