--- src/pages/deposit/index.css
@@ -1,21 +1,0 @@
-#deposit {
-    width: 600px;
-    margin: 90px;
-    padding: 20px;
-    margin-left: 380px;
-   margin-top: 140px;
-  }
-#deposit_stk{
-    width: 410px;
-    left:   1px
-}
-#deposit_amount{
-    width: 411px;
-    left:   54px
-}
-#deposit_btn{
-    left:463px
-}
-.mainBody{
-    border-top: 0.01px solid silver;
-}
